@@ -39,30 +39,30 @@
 		"@babylonjs/gui": "^4.2.0",
 		"@babylonjs/loaders": "^4.2.0",
 		"@lowesinnovationlab/viewer-assets": "^0.0.4",
-		"@types/react": "17.0.11",
-		"@types/react-dom": "17.0.7",
-		"@typescript-eslint/eslint-plugin": "^4.26.1",
-		"@typescript-eslint/parser": "^4.26.1",
-		"chokidar": "^3.5.1",
-		"clean-webpack-plugin": "^4.0.0-alpha.0",
-		"concurrently": "^6.2.1",
-		"copy-webpack-plugin": "^9.0.0",
+		"@types/react": "17.0.37",
+		"@types/react-dom": "17.0.11",
+		"@typescript-eslint/eslint-plugin": "^5.6.0",
+		"@typescript-eslint/parser": "^5.6.0",
+		"chokidar": "^3.5.2",
+		"clean-webpack-plugin": "^4.0.0",
+		"concurrently": "^6.4.0",
+		"copy-webpack-plugin": "^10.1.0",
 		"eslint": "^7.32.0",
-		"eslint-config-prettier": "^6.13.0",
-		"eslint-plugin-prettier": "^3.4.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-prettier": "^4.0.0",
 		"file-loader": "^6.2.0",
-		"fork-ts-checker-webpack-plugin": "^6.2.13",
-		"html-webpack-plugin": "^5.3.1",
-		"lint-staged": "^10.5.4",
-		"lit": "^2.0.0-rc.2",
-		"prettier": "^2.3.1",
+		"fork-ts-checker-webpack-plugin": "^6.5.0",
+		"html-webpack-plugin": "^5.5.0",
+		"lint-staged": "^12.1.2",
+		"lit": "^2.0.2",
+		"prettier": "^2.5.1",
 		"source-map-loader": "^3.0.0",
 		"symlink-dir": "^5.0.1",
-		"ts-loader": "^9.2.3",
+		"ts-loader": "^9.2.6",
 		"typescript": "4.0.2",
-		"webpack": "^5.38.1",
-		"webpack-cli": "^4.7.2",
-		"webpack-dev-server": "^3.11.2",
+		"webpack": "^5.65.0",
+		"webpack-cli": "^4.9.1",
+		"webpack-dev-server": "^4.6.0",
 		"webpack-merge": "^5.8.0"
 	},
 	"lint-staged": {
