@@ -6,6 +6,7 @@
 		"node": ">=10.0.0"
 	},
 	"scripts": {
+		"preinstall": "npx npm-force-resolutions",
 		"start": "react-scripts --max_old_space_size=4096 start",
 		"build:watch": "npm run start",
 		"build": "react-scripts --max_old_space_size=4096 build",
@@ -23,32 +24,32 @@
 	},
 	"dependencies": {
 		"@lowesinnovationlab/product-viewer": "^0.0.5",
-		"@material-ui/core": "^4.11.4",
+		"@material-ui/core": "^4.12.3",
 		"@material-ui/icons": "^4.11.2",
-		"@testing-library/jest-dom": "^4.2.4",
-		"@testing-library/react": "^9.5.0",
-		"@testing-library/user-event": "^7.2.1",
-		"@types/jest": "^24.9.1",
-		"@types/node": "^12.20.15",
+		"@testing-library/jest-dom": "^5.16.1",
+		"@testing-library/react": "^12.1.2",
+		"@testing-library/user-event": "^13.5.0",
+		"@types/jest": "^27.0.3",
+		"@types/node": "^16.11.12",
 		"@types/prismjs": "^1.16.6",
-		"@types/react": "^16.14.8",
-		"@types/react-dom": "^16.9.13",
-		"@types/react-redux": "^7.1.16",
-		"@types/react-router-dom": "^5.1.7",
-		"@types/styled-components": "^5.1.10",
-		"prismjs": "^1.24.1",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
-		"react-redux": "^7.2.4",
-		"react-router-dom": "^5.2.0",
+		"@types/react": "^17.0.37",
+		"@types/react-dom": "^17.0.11",
+		"@types/react-redux": "^7.1.20",
+		"@types/react-router-dom": "^5.3.2",
+		"@types/styled-components": "^5.1.17",
+		"prismjs": "^1.25.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"react-redux": "^7.2.6",
+		"react-router-dom": "^6.1.0",
 		"react-scripts": "^4.0.3",
 		"react-split-pane": "^0.1.92",
-		"redux": "^4.1.0",
+		"redux": "^4.1.2",
 		"redux-devtools-extension": "^2.13.9",
-		"redux-thunk": "^2.3.0",
-		"sass": "^1.34.1",
-		"styled-components": "^5.3.0",
-		"typescript": "^3.9.9"
+		"redux-thunk": "^2.4.1",
+		"sass": "^1.45.0",
+		"styled-components": "^5.3.3",
+		"typescript": "^4.0.2"
 	},
 	"eslintConfig": {
 		"extends": "react-app"
@@ -73,15 +74,23 @@
 	},
 	"devDependencies": {
 		"@lowesinnovationlab/viewer-assets": "^0.0.4",
-		"@typescript-eslint/eslint-plugin": "^4.26.1",
-		"@typescript-eslint/parser": "^4.26.1",
+		"@typescript-eslint/eslint-plugin": "^5.6.0",
+		"@typescript-eslint/parser": "^5.6.0",
 		"eslint": "^7.32.0",
-		"eslint-config-prettier": "^6.13.0",
-		"eslint-plugin-prettier": "^3.4.0",
-		"fast-sass-loader": "^1.5.0",
-		"is-svg": "^4.3.1",
-		"lint-staged": "^10.5.4",
-		"prettier": "^2.3.1",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-prettier": "^4.0.0",
+		"fast-sass-loader": "^2.0.1",
+		"is-svg": "^4.3.2",
+		"lint-staged": "^12.1.2",
+		"prettier": "^2.5.1",
 		"symlink-dir": "^5.0.1"
+	},
+	"resolutions": {
+		"ansi-html": "https://registry.yarnpkg.com/ansi-html-community/-/ansi-html-community-0.0.8.tgz",
+		"ansi-regex": "5.0.1",
+		"browserslist": "4.18.1",
+		"glob-parent": "6.0.2",
+		"immer": "9.0.6",
+		"nth-check": "2.0.1"
 	}
 }
